Refactor SpaceCardGame into smaller components
